requirements.txt
# LakebaseOps Platform Dependencies
# Core (required)
databricks-sdk>=0.81.0
psycopg[binary]>=3.0
requests>=2.31.0

# PySpark (available in Databricks runtime, install locally for dev)
# pyspark>=3.5.0

# Testing
pytest>=7.0
pytest-asyncio>=0.21

# E2E testing (Aurora provisioning)
boto3>=1.34

# Optional: AI features (V2)
# databricks-genai-inference>=0.1.0
